Safari-609.1.11.1/LayoutTests/media/modern-media-controls/start-support/start-support-autoplay-expected.txt

Testing the StartSupport behavior with autoplay.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Media is playing
PASS mediaController.controls.showsStartButton is false

PASS successfullyParsed is true

TEST COMPLETE
